+++ src/pages/NossosServicos.tsx
@@ -1,5 +1,5 @@
 import Header from "../components/Header";
-import SecaoNossosServicos from "../components/SecaoNossosServicos/SecaoNossosServicos";
+import SecaoNossosServicos from "../sections/SecaoNossosServicos";
 
 export default function NossosServicos() {
   return (
